Stop reading the previous sentence back as this one's context
Six findings from an external review, each verified against the source before
being acted on rather than taken on report. All six were real. The suite it
could not run is green at 351 tests, both flavors, so nothing was already
failing.

The one that mattered: refreshContextFromCursor computed whether the cursor
sat at a sentence start, then took the preceding word by scanning straight
across the boundary it had just found. Two correct lines, four apart,
contradicting each other. After "Hello. " the separator set an empty context
and the strip showed openers; the cursor update arrived milliseconds later,
overwrote it with "hello", and the openers were replaced by continuations --
visible as a flicker. The next word committed was then learned as following
"hello", across the full stop, which is precisely what the atSentenceStart
field was introduced to prevent; its own doc comment says so. The opener model
was very nearly unreachable, since an empty context now required no word
characters at all in the preceding text and a field that empty is short-
circuited earlier.

Moved into SentenceContext, pure and tested, because it is a rule about text
rather than about a keyboard, and nothing in the suite had touched either line.
Cutting at the boundary also fixes a case the report did not mention: one word
into a new sentence, prevWord2 was the last word of the previous one.

The rest, all confirmed:

The n-gram decay guards on table size but halving only removes a context whose
counts were all 1, so on a mature model a pass could leave the size unchanged
and the next committed word would sweep six thousand contexts again, on the
main thread. Halved until actually under the cap now, once, with a bound.

The split guard multiplied an Int frequency by 150. Checked the shipped
dictionaries rather than assuming: "that" at 10.2M is the most frequent word of
four or more letters anywhere, against an overflow at 14.3M -- unreachable, by
40%, and with no bound at all on an imported dictionary.

Autocorrect skipped any word with a capital after the first letter, which is
right for iPhone and wrong for caps lock: TEH stayed TEH. All-capitals is now
corrected, and acronyms need no special case because a real one is in the
dictionary and returns before this.

Each engine's warm executor is released on destroy -- my own code from earlier
this week, and the spell checker's teardown reasoned about UserData's executor
while missing it.

The learned-word files go through a temporary file and a rename, since two
UserData instances write them and writeText truncates before writing.

+++ app/src/main/java/com/rimboard/keyboard/engine/Dictionary.kt
@@ -495,7 +495,14 @@ class Dictionary(
             if (fa < floorFor(a) || fb < floorFor(b)) continue
             // An attested word is only a missing space if it is overwhelmingly
             // rarer than the two words it would become.
-            if (typedFreq > 0 && minOf(fa, fb) < typedFreq * SPLIT_DOMINANCE) continue
+            // Long: the product overflows Int above a typed frequency of about
+            // 14.3 million, and an overflowed product goes negative, which
+            // makes the comparison false and *skips the guard* — inverting it
+            // for the commonest words in the corpus. Unreachable today (the
+            // most frequent word of four or more letters in any shipped
+            // dictionary is "that" at 10.2M) but only by 40%, and an imported
+            // dictionary carries no bound at all.
+            if (typedFreq > 0 && minOf(fa, fb) < typedFreq.toLong() * SPLIT_DOMINANCE) continue
             // Both halves count, so a split into two common words beats one
             // into a common word and a rare one.
             val score = ln((fa + 1).toDouble()) + ln((fb + 1).toDouble())
